eth: split relay-fetch code out of handler_wit.go
Pure extraction, no functional change: triggerRelayFetch,
fetchAndVerifyWitness, fetchWitnessPages, and requestWitnessPage move
to a new eth/handler_wit_relay_fetch.go, pairing with the existing
eth/handler_wit_relay_fetch_test.go.

Fixes the CI quality-gate failure: handler_wit.go had grown to 812
lines, over the 800-line file-size threshold. Verified with diffguard
using CI's exact thresholds that the Code Sizes check now passes.
